Refactor AI review application logic and enhance utility functions
- Moved `craftReviewPrompt` and `parseAIResponse` functions to a new utility file (`ai-review-utils.ts`) for better organization and reusability.
- Updated the AI review endpoint to utilize the new utility functions, improving code clarity and maintainability.
- Implemented enhanced error handling and response validation in the AI review process, ensuring robust feedback and user-friendly error messages.
- Optimized the prompt generation and AI response parsing logic to streamline the review process and improve overall efficiency.

These changes improve the structure of the codebase, enhance the user experience, and ensure more reliable AI review functionality.
