+++ tests/unit/ai-review.test.ts
@@ -1,5 +1,5 @@
 
-import { parseAIResponse, craftReviewPrompt } from '@/app/api/ai/review-application/route';
+import { parseAIResponse, craftReviewPrompt } from '@/lib/ai-review-utils';
 
 // Mock data for testing
 const mockRequirement = {
